# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Passport API
  description: |
    **Sign in with Google, but for Agents.**
    
    Agent Passport is an OAuth-like identity layer for AI agents. It provides:
    - Agent registration with Ed25519 public keys
    - Challenge-response authentication
    - Short-lived JWT identity tokens
    - Token verification for app builders
    - Rate limiting and risk scoring
    
    ## Quick Links
    - [Integration Guide](./INTEGRATION.md)
    - [Smoke Test](./SMOKE_TEST.md)
    - [Security](./SECURITY.md)
  version: 0.1.0
  contact:
    name: Agent Passport
    url: https://github.com/zerobase-labs/agent-passport
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://passport-api.onrender.com
    description: Production API (Render)
  - url: http://localhost:3000
    description: Local development

tags:
  - name: Health
    description: Health check endpoints
  - name: Agents
    description: Agent registration and authentication
  - name: Tokens
    description: Token verification for app builders
  - name: JWKS
    description: JSON Web Key Set for token verification

paths:
  /healthz:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns API health status and version
      operationId: healthCheck
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: ok
                version: 0.1.0
                timestamp: '2026-02-01T12:00:00.000Z'

  /agents/register:
    post:
      tags:
        - Agents
      summary: Register a new agent
      description: |
        Register a new AI agent with an Ed25519 public key.
        
        The agent must generate an Ed25519 keypair and provide the base64-encoded public key.
        The private key never leaves the agent.
      operationId: registerAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAgentRequest'
            example:
              publicKey: "MCowBQYDK2VwAyEA..."
              name: "My AI Assistant"
              metadata:
                model: "gpt-4"
                capabilities: ["chat", "code"]
      responses:
        '201':
          description: Agent registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterAgentResponse'
        '400':
          description: Invalid request (bad public key format, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Agent with this public key already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /agents/{agentId}/challenge:
    post:
      tags:
        - Agents
      summary: Request authentication challenge
      description: |
        Request a challenge nonce for authentication.
        
        The agent must sign this nonce with their private key and submit it
        to the identity-token endpoint within 5 minutes.
        
        **Rate Limited:** 60 requests/minute per agent, 120 requests/minute per IP
      operationId: getChallenge
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Challenge issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeResponse'
              example:
                nonce: "abc123..."
                expiresAt: "2026-02-01T12:05:00.000Z"
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /agents/{agentId}/identity-token:
    post:
      tags:
        - Agents
      summary: Exchange signed challenge for identity token
      description: |
        Submit a signed challenge to receive a short-lived JWT identity token.
        
        The signature must be a valid Ed25519 signature of the challenge nonce
        using the agent's registered private key.
        
        **Rate Limited:** 30 requests/minute per agent, 60 requests/minute per IP
      operationId: getIdentityToken
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdentityTokenRequest'
            example:
              nonce: "abc123..."
              signature: "MEUCIQDk..."
      responses:
        '200':
          description: Identity token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityTokenResponse'
        '400':
          description: Invalid signature or expired nonce
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /tokens/verify:
    post:
      tags:
        - Tokens
      summary: Verify an agent identity token
      description: |
        Verify an agent's identity token and get agent information.
        
        **Requires App Authentication:** Include `X-App-Id` and `X-App-Key` headers.
        
        **Rate Limited:** 600 requests/minute per app, 120 requests/minute per IP
        
        Returns the agent's identity claims plus risk assessment.
      operationId: verifyToken
      security:
        - appAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyTokenRequest'
            example:
              token: "eyJhbGciOiJFZERTQSIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Token verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyTokenResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenErrorResponse'
        '401':
          description: Invalid app credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /tokens/revoke:
    post:
      tags:
        - Tokens
      summary: Revoke an agent identity token
      description: |
        Revoke a JWT identity token before it expires. The token's JTI is added
        to a Redis blocklist with a TTL matching the token's remaining lifetime.
        Subsequent verify calls for this token will return `valid: false` with
        reason `token_revoked`.
        
        **Requires App Authentication:** Include `X-App-Id` and `X-App-Key` headers.
        
        Revoking an already-expired token is idempotent and returns success.
      operationId: revokeToken
      security:
        - appAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeTokenRequest'
            example:
              token: "eyJhbGciOiJFZERTQSIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Token revoked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeTokenResponse'
        '401':
          description: Invalid app credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Token could not be revoked (e.g., not a valid JWT)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /tokens/introspect:
    post:
      tags:
        - Tokens
      summary: Introspect a token (RFC 7662)
      description: |
        Introspect a token per RFC 7662 OAuth 2.0 Token Introspection.
        
        **Requires App Authentication:** Include `X-App-Id` and `X-App-Key` headers.
        
        Returns whether the token is active and its claims.
      operationId: introspectToken
      security:
        - appAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntrospectRequest'
            example:
              token: "eyJhbGciOiJFZERTQSIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Introspection result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntrospectResponse'
        '401':
          description: Invalid app credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /.well-known/jwks.json:
    get:
      tags:
        - JWKS
      summary: Get JSON Web Key Set
      description: |
        Returns the public keys used to sign identity tokens.
        
        Apps can use this to verify tokens locally without calling the verify endpoint.
      operationId: getJwks
      responses:
        '200':
          description: JWKS response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwksResponse'
              example:
                keys:
                  - kty: OKP
                    crv: Ed25519
                    x: "11qYAYKxCrfVS_7TyWQHOg7hcvPapiMlrwIaaPcHURo"
                    kid: "passport-ed25519-001"
                    use: sig
                    alg: EdDSA

components:
  securitySchemes:
    appAuth:
      type: apiKey
      in: header
      name: X-App-Id
      description: |
        App authentication requires two headers:
        - `X-App-Id`: Your app's ID
        - `X-App-Key`: Your app's secret key (plaintext)

  parameters:
    AgentId:
      name: agentId
      in: path
      required: true
      description: The unique agent identifier
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"

  headers:
    X-RateLimit-Limit:
      description: The maximum number of requests allowed in the window
      schema:
        type: integer
      example: 60
    X-RateLimit-Remaining:
      description: The number of requests remaining in the window
      schema:
        type: integer
      example: 45
    X-RateLimit-Reset:
      description: Unix timestamp when the rate limit resets
      schema:
        type: integer
      example: 1738411200
    Retry-After:
      description: Seconds to wait before retrying
      schema:
        type: integer
      example: 30

  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - version
        - timestamp
      properties:
        status:
          type: string
          enum: [ok]
        version:
          type: string
        timestamp:
          type: string
          format: date-time

    RegisterAgentRequest:
      type: object
      required:
        - publicKey
      properties:
        publicKey:
          type: string
          description: Base64-encoded Ed25519 public key (SPKI format)
          pattern: "^[A-Za-z0-9+/=]+$"
        name:
          type: string
          maxLength: 255
          description: Human-readable agent name
        metadata:
          type: object
          description: Optional metadata (model, capabilities, etc.)
          additionalProperties: true

    RegisterAgentResponse:
      type: object
      required:
        - agentId
        - publicKey
        - status
        - createdAt
      properties:
        agentId:
          type: string
          format: uuid
        publicKey:
          type: string
        name:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, suspended]
        metadata:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time

    ChallengeResponse:
      type: object
      required:
        - nonce
        - expiresAt
      properties:
        nonce:
          type: string
          description: Random nonce to sign (hex or base64)
        expiresAt:
          type: string
          format: date-time
          description: When the challenge expires (5 minutes)

    IdentityTokenRequest:
      type: object
      required:
        - nonce
        - signature
      properties:
        nonce:
          type: string
          description: The challenge nonce that was signed
        signature:
          type: string
          description: Base64-encoded Ed25519 signature of the nonce

    IdentityTokenResponse:
      type: object
      required:
        - token
        - expiresAt
      properties:
        token:
          type: string
          description: JWT identity token (EdDSA signed)
        expiresAt:
          type: string
          format: date-time
          description: When the token expires (60 minutes)

    VerifyTokenRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: The JWT identity token to verify

    VerifyTokenResponse:
      type: object
      required:
        - valid
        - agentId
        - publicKey
        - status
        - iat
        - exp
        - risk
      properties:
        valid:
          type: boolean
          enum: [true]
        agentId:
          type: string
          format: uuid
        publicKey:
          type: string
        name:
          type: string
          nullable: true
        status:
          type: string
          enum: [active]
        iat:
          type: integer
          description: Token issued-at (Unix timestamp)
        exp:
          type: integer
          description: Token expiration (Unix timestamp)
        risk:
          $ref: '#/components/schemas/RiskAssessment'

    RiskAssessment:
      type: object
      required:
        - score
        - recommendedAction
        - reasons
      properties:
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: Risk score (0=low risk, 100=high risk)
        recommendedAction:
          type: string
          enum: [allow, throttle, block]
          description: Recommended action based on risk score
        reasons:
          type: array
          items:
            type: string
          description: Human-readable risk factors

    TokenErrorResponse:
      type: object
      required:
        - valid
        - reason
      properties:
        valid:
          type: boolean
          enum: [false]
        reason:
          type: string
          enum:
            - invalid_signature
            - expired
            - token_revoked
            - agent_suspended
            - agent_not_found
            - invalid_token
            - internal_error
          description: Why the token is invalid

    RevokeTokenRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: The JWT identity token to revoke

    RevokeTokenResponse:
      type: object
      required:
        - revoked
        - jti
        - expiresAt
      properties:
        revoked:
          type: boolean
          enum: [true]
        jti:
          type: string
          description: The JWT ID that was revoked
        expiresAt:
          type: string
          format: date-time
          description: When the blocklist entry expires (matches token expiry)

    IntrospectRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: The token to introspect

    IntrospectResponse:
      type: object
      required:
        - active
      properties:
        active:
          type: boolean
          description: Whether the token is active
        sub:
          type: string
          description: Subject (agent ID)
        iat:
          type: integer
          description: Issued-at timestamp
        exp:
          type: integer
          description: Expiration timestamp
        iss:
          type: string
          description: Issuer
        agent_name:
          type: string
          nullable: true
        agent_status:
          type: string

    JwksResponse:
      type: object
      required:
        - keys
      properties:
        keys:
          type: array
          items:
            type: object
            properties:
              kty:
                type: string
                enum: [OKP]
              crv:
                type: string
                enum: [Ed25519]
              x:
                type: string
                description: Base64url-encoded public key
              kid:
                type: string
              use:
                type: string
                enum: [sig]
              alg:
                type: string
                enum: [EdDSA]

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string